# docs/openapi_admin.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "管理员端接口",
    "description": "管理员端的相关api。本地开发环境需要删除/api前缀再进行请求，服务器生产环境配置了nginx路由，直接请求即可。最终版。",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/admin/post/{id}": {
      "get": {
        "summary": "获取帖子详情",
        "deprecated": false,
        "description": "点击列表中的查看按钮调用。",
        "tags": [
          "ContentAudit"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "693fae15a7f63f5102d5084a",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "author": {
                          "type": "object",
                          "properties": {
                            "userId": {
                              "type": "integer"
                            },
                            "nickname": {
                              "type": "string"
                            },
                            "avatar": {
                              "type": "string"
                            },
                            "email": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "userId",
                            "nickname",
                            "avatar",
                            "email"
                          ]
                        },
                        "title": {
                          "type": "string"
                        },
                        "content": {
                          "type": "string"
                        },
                        "type": {
                          "type": "integer"
                        },
                        "cover": {
                          "type": "string"
                        },
                        "images": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "video": {
                          "type": "string"
                        },
                        "likeCount": {
                          "type": "integer"
                        },
                        "collectCount": {
                          "type": "integer"
                        },
                        "commentCount": {
                          "type": "integer"
                        },
                        "isLiked": {
                          "type": "null"
                        },
                        "isCollected": {
                          "type": "null"
                        },
                        "isFollowed": {
                          "type": "null"
                        },
                        "ratingAverage": {
                          "type": "integer"
                        },
                        "ratingCount": {
                          "type": "integer"
                        },
                        "myScore": {
                          "type": "null"
                        },
                        "createdAt": {
                          "type": "string"
                        },
                        "tags": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "coverWidth": {
                          "type": "integer"
                        },
                        "coverHeight": {
                          "type": "integer"
                        }
                      },
                      "required": [
                        "id",
                        "author",
                        "title",
                        "content",
                        "type",
                        "cover",
                        "images",
                        "video",
                        "likeCount",
                        "collectCount",
                        "commentCount",
                        "isLiked",
                        "isCollected",
                        "isFollowed",
                        "ratingAverage",
                        "ratingCount",
                        "myScore",
                        "createdAt",
                        "tags",
                        "coverWidth",
                        "coverHeight"
                      ]
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ]
                },
                "example": {
                  "code": 200,
                  "message": "操作成功",
                  "data": {
                    "id": "693fae15a7f63f5102d5084a",
                    "author": {
                      "userId": "9",
                      "nickname": "映记测试账号",
                      "avatar": "https://afternoon3-rednote.oss-cn-shenzhen.aliyuncs.com/default_avatar.jpg",
                      "email": "test5@test.com"
                    },
                    "title": "测试视频接口",
                    "content": "测试视频接口测试视频接口测试视频接口测试视频接口测试视频接口测试视频接口测试视频接口测试视频接口测试视频接口测试视频接口测试视频接口测试视频接口测试视频接口测试视频接口测试视频接口测试视频接口测试视频接口",
                    "type": 1,
                    "cover": "https://afternoon3-rednote.oss-cn-shenzhen.aliyuncs.com/erhai-lake.mp4?x-oss-process=video/snapshot,t_1000,f_jpg,w_0,h_0,m_fast",
                    "images": [
                      "...",
                      "..."
                    ],
                    "video": "https://afternoon3-rednote.oss-cn-shenzhen.aliyuncs.com/erhai-lake.mp4",
                    "likeCount": 1,
                    "collectCount": 0,
                    "commentCount": 1,
                    "isLiked": null,
                    "isCollected": null,
                    "isFollowed": null,
                    "ratingAverage": 0,
                    "ratingCount": 0,
                    "myScore": null,
                    "createdAt": "2025-12-15T14:43:33.743",
                    "tags": [
                      "洱海",
                      "风景"
                    ],
                    "coverWidth": 1280,
                    "coverHeight": 720
                  }
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "未找到资源响应",
                  "type": "object",
                  "properties": {
                    "code": {
                      "title": "状态码",
                      "type": "integer",
                      "example": 404
                    },
                    "msg": {
                      "title": "错误提示",
                      "type": "string",
                      "example": "帖子不存在或已被删除"
                    },
                    "data": {
                      "title": "数据",
                      "type": "null"
                    }
                  },
                  "required": [
                    "code",
                    "msg"
                  ]
                },
                "example": {
                  "code": 404,
                  "msg": "帖子不存在或已被删除",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
...
    "/api/admin/post/{postId}": {
      "delete": {
        "summary": "管理员删除帖子",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "693c5c9951cd4f43d5167f5e",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reason",
            "in": "query",
            "description": "",
            "required": false,
            "example": "测试删除帖子",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "jwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [],
  "security": [
    {
      "jwt": []
    }
  ]
}